- serial device moved to separate class definition
